Modularize FileEditorTool internals
Extract schema, content transforms, view handlers, and undo-history management into focused fileEditor modules while preserving existing command behavior.
